code_final/food101/requirements.txt
cycler
decorator
imageio
joblib
kiwisolver
matplotlib
netcal
networkx
numpy
Pillow
pyparsing
pandas
numpy
plotly
torchvision
python-dateutil
PyWavelets
scikit-image
scikit-learn
scipy
six
sklearn
threadpoolctl
tifffile
tqdm
typing-extensions
statsmodels
KDEpy
jax
jaxlib
autograd
torch <= 1.3
torchvision